Use a proper buffer name in a prepare_notification() test.
Buffers in WeeChat are, in fact, buffer identifiers (e.g. '0x456b34). To obtain
the short name or name of the buffer, one has to use

    weechat.get_buffer_string(ID, 'short_name')

or

    weechat.get_buffer_string(ID, 'name')

respectively. This is how it is done in the plugin. The tests, however, assume
that buffer ID and name are the same thing, which is false. This commit fixes
that by setting buffer's short name to '#buffer'.