Improve the detection of nicks from data passed by WeeChat.
Originally, the nick was obtained from the prefix. However, the prefix is not
always the nick. Now, the nick is first tried to be obtained from a tag of the
form 'nick_XYZ', where XYZ is the nick that sent the message. If this fails,
the nick is obtained from the prefix (fallback).