Consider >channel and #channel to be equal names.
The reason is that the https://github.com/rawdigits/wee-slack plugin replaces
'#' with '>' to indicate that someone in the buffer is typing. This fixes the
behavior of several configuration options (e.g.
'notify_on_all_messages_in_buffers') when weechat_notify_send is used together
with the wee_slack plugin.

Note that this is only needed to be done for the short name. Indeed,
the full name always stays unchanged.